Fix crash when ArrayLength is zero
When loading in the default values for some OSL files, the array length is zero, but when then calling array.reserve, it crashes. I've added in a simple check to make sure that the ArrayLength is greater than zero, which fixes this issue. The reserve function should in theory already be able to handle this, but it's probably the cast to size_t that causes the issue